postales/app/requirements.txt
mysql-connector-python
python-dotenv
pandas
fastapi
uvicorn[standard]